mFix Auth

+++ app/Http/Controllers/API/Auth/PasswordResetController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API\Auth;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Mail;
@@ -55,7 +56,41 @@ class PasswordResetController extends Controller
     // POST api/account/check-token
     public function checkToken(Request $request)
     {
-        $credential = DB::table('password_reset_tokens')->where('email', $request->email)->delete();
-        return ApiResponse(true,  null, Response::HTTP_OK, 'Token hợp lệ');
+        try {
+            $validator = Validator::make(
+                $request->all(),
+                [
+                    'password' => 'required|string|min:8|max:10',
+                ],
+                [
+                    'required' => 'Trường :attribute không được để trống',
+                    'string' => 'Trường :attribute phải là chuỗi kí tự',
+                    'min' => 'Trường :attribute có độ dài tối thiểu :min kí tự',
+                    'max' => 'Trường :attribute có độ dài tối đa là :max kí tự',
+                ],
+                [
+                    'password' => 'Mật khẩu mới',
+                ]
+            );
+    
+            if ($validator->fails()) {
+                $credential = DB::table('password_reset_tokens')->where('email', $request->email)->delete();
+                return ApiResponse(false, null, Response::HTTP_BAD_REQUEST, ['errors' => $validator->errors()]);
+            }
+    
+            $user = User::where('email', $request->email)->first();
+            if (!$user) {
+                $credential = DB::table('password_reset_tokens')->where('email', $request->email)->delete();
+                return ApiResponse(true,  null, Response::HTTP_OK, 'Email không tồn tại');
+            }
+    
+            $user->password = Hash::make(trim($request->password));
+            $user->save();
+    
+            $credential = DB::table('password_reset_tokens')->where('email', $request->email)->delete();
+            return ApiResponse(true, $user, Response::HTTP_OK, 'Reset password thành công, vui lòng đăng nhập lại để sử dụng dịch vụ');
+        } catch (\Exception $e) {
+            return ApiResponse(false, null, Response::HTTP_BAD_REQUEST, $e->getMessage());
+        }
     }
 }
